Add a more efficient PathExtractor Implementation
The existing implementation loops through every potential search
path at each step of matching. And therefore has exponential time
complexity as the number of extracted fields grows.

This change adds a more efficient PathExtractor implementation that
supports a narrower set of SearchPaths and combinations thereof. But
avoids the exponential time cost.

In the benchmark testing on my machine, I found it to be significantly
faster for the "full" extraction and slightly faster for the "partial"
extraction. With the throughput score from the benchmarking about
6.5:3.8 and 48:42. I also found it to be significantly faster (roughly 3:5)
for extracting most of the fields out of binary Ion dataset of about
20k records, each with about 15 distinct elements, some nested.

It is built with a key assumption that does not hold for the existing
implementation: that users may not register paths that would result
in a single element matching more than one terminal matcher.

I searched for Amazon internal usages and believe that the supported
usage patterns cover any and all usage I could find. I found no usage
of Annotations not on the Top-Level-Values themselves.