Add Type Check to Strict Extractor (#34)
* Add Type Check to Strict Extractor

Search paths on field names, indexed elements or wildcards imply an
expectation for container types. Before this change, a search path
looking for a field would simply not be invoked if the data was not
a struct. With this change, users can elect to have an Exception thrown
when data does not meet the implied type expectations of their paths.

The type check is performed after any callback for the parent itself:
that is not known to be mistyped by the paths.

It is only implemented for the "strict" matcher and the change to the
API reflects that.

* Add Test for null.struct